% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{A 1,000 sample population with physical information.}
\format{
A data frame with 1,000 rows and 5 variables:
\describe{
\item{ID}{Identifier, ID#}
\item{gender}{The gender of the participant}
\item{height}{The height of the person}
\item{weight}{The weight of the person}
\item{phys}{How intense is this person's physical activities}
}
}
\usage{
dat
}
\description{
A data set containing the gender, height, weight and physical activities of the 1,000 people.
}
\keyword{datasets}
