% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATProject.R
\name{StatTest}
\alias{StatTest}
\title{Generate output of a statistical test}
\usage{
StatTest(test)
}
\arguments{
\item{test}{text string; This is one of the test that should be implemented, such as lm, ttest or chitest.}
}
\value{
A list of test output information, yet less user-friendly as this is not expected to be viewed by general users. However, the default print function would print the user-friendly version of the output.
}
\description{
Generate output of a statistical test
}
\examples{
StatTest(test="chitest")
}
